Manage duration as a float
Double type is not constructed the same way in C++ and Java, and does
not exist in python. Our bindings of 'C++ double' could create issues,
and since a fps with a simple precision is enough, we decided to return
fps as float.

+++ src/AvTranscoder/mediaProperty/FileProperties.cpp
@@ -157,7 +157,7 @@ double FileProperties::getStartTime() const
 	return  1.0 * (unsigned int)_avFormatContext->start_time / AV_TIME_BASE;
 }
 
-double FileProperties::getDuration() const
+float FileProperties::getDuration() const
 {
 	if( ! _avFormatContext )
 		throw std::runtime_error( "unknown format context" );
