+++ src/AvTranscoder/stream/InputStream.hpp
@@ -26,7 +26,7 @@ public:
 
 	size_t getStreamIndex() const { return _streamIndex; }
 	/// Get duration of the stream, in seconds
-	double getDuration() const;
+	float getDuration() const;
 	AVMediaType getStreamType() const;
 
 	VideoCodec& getVideoCodec();
